Optimize region highlighter by finding non-matching lines faster
Our regex engine uses the virtual machine approach, see
https://swtch.com/~rsc/regexp/regexp2.html
It is very robust but still does quite a bit of work if the pattern
prefix matches in many places, like for this regex that tries to
match the start of a URL:

	[a-z]+://

For every lowercase word in the buffer, we spawn a bunch of VM threads
only to throw most of them away later, since most words don't contain
"://".

Since the region highlighter works line-by-line, we can skip lines
that do not match the highlighter regex.  Most highlighters have some
literal parts. Extract literal parts that must always match. Before
performing a full regex match, check if the line contains that literal
part.  This is much faster in practice. Of course it's a heuristic,
so it can also cause slow downs.  Note that this optimization only
applies to the region highlighter; it would be less effective for
searches that are not limited to a single line.

The actual regex used by markdown.kak is this one:

	add-highlighter shared/markdown/inline/url region -recurse \( ([a-z]+://|(mailto|magnet|xmpp):) (?!\\).(?=\))|\s fill link

which takes almost 0.5 seconds to apply to the [5MB Markdown file].
With this optimization it only takes 0.005 seconds. The speedup on
the entire file is also solid:

	$ HOME=$PWD hyperfine -w 1 ./kak.opt.{old,new}' big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy'
	Benchmark 1: ./kak.opt.old big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy
	  Time (mean ± σ):      6.478 s ±  0.096 s    [User: 6.442 s, System: 0.030 s]
	  Range (min … max):    6.297 s …  6.570 s    10 runs
	 
	Benchmark 2: ./kak.opt.new big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy
	  Time (mean ± σ):      1.966 s ±  0.038 s    [User: 1.940 s, System: 0.029 s]
	  Range (min … max):    1.915 s …  2.021 s    10 runs
	 
	Summary
	  './kak.opt.new big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy' ran
	    3.30 ± 0.08 times faster than './kak.opt.old big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy'

[5MB Markdown file]: https://github.com/mawww/kakoune/issues/4685#issuecomment-1208129806
Here is the script to create that file:

	: >big_markdown.md
	for i in $(seq 10000); do
	  cat <<EOF >> big_markdown.md
	# Title
	## Subtitle
	Some notes about things.

	Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
	incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
	nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
	Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore
	eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt
	in culpa qui officia deserunt mollit anim id est laborum

	* A list
	  * With a sub-item
	* Item 2
	* Item 3
	* Item 4
	* Item 5
	  * Item 5.1

	EOF
	done