Allow empty region end in region highlighter
Some of our highlighters abuse the regions highlighter by using one
of these patterns as region end:

	'\K'
	'.\K'
	(?!\\).(?=>)|\n
	(?!\\).(?=\))|\s
	\b

These highlighters should be regex highlighters (because they have
no children) but are probably required to be region highlighters to
compete with other region highlighters.

A region highlighter works by first computing *all* matches of both
region start and region end (which faciliates caching).  Unfortunately
this means that these oddball region-end-patterns create loads of
matches, most of which are useless.

Let's standardize these hacks by allowing the region end to be empty.
This allows us to match the entire region in the region begin pattern,
and do no extra work if the region end is empty.

It's still hacky (the region highlighter feels like the wrong tool
for some of these cases) but I don't think we have a good alternative
today.

On the [5MB Markdown file] this gives a decent speedup for the initial
highlighting pass:

	$ HOME=$PWD hyperfine -w 1 'git checkout HEAD'{~,}' -- :/rc/filetype/markdown.kak && ./kak.opt big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy'
	Benchmark 1: git checkout HEAD~ -- :/rc/filetype/markdown.kak && ./kak.opt big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy
	  Time (mean ± σ):      1.969 s ±  0.033 s    [User: 1.935 s, System: 0.035 s]
	  Range (min … max):    1.906 s …  2.036 s    10 runs
	 
	Benchmark 2: git checkout HEAD -- :/rc/filetype/markdown.kak && ./kak.opt big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy
	  Time (mean ± σ):      1.111 s ±  0.022 s    [User: 1.093 s, System: 0.020 s]
	  Range (min … max):    1.079 s …  1.145 s    10 runs
	 
	Summary
	  'git checkout HEAD -- :/rc/filetype/markdown.kak && ./kak.opt big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy' ran
	    1.77 ± 0.05 times faster than 'git checkout HEAD~ -- :/rc/filetype/markdown.kak && ./kak.opt big_markdown.md -e "hook global NormalIdle .* quit" -ui dummy'

(The numbers depend on a other optimizations I'm using.)

---

Regarding markdown.kak :

To avoid the empty-region-end hack I considered using regex
highlighters (instead of region highlighters) for Markdown
URLs.  However, we probably want to keep highlighting bare
URLs (without enclosing <>).  This violates the [CommonMark
spec](https://spec.commonmark.org/0.29/#example-607) but both GitHub
Markdown and editors like VSCode and Emacs highlight bare URLs here.

The problem with highlighting bare URLs is that
they may contain Markdown metacharacters, see
https://spec.commonmark.org/dingus/?text=https%3A%2F%2Fexample.com%2F(1)_2_(3)

So unless we keep using a region highlighter, it will be very difficult
to avoid spurious Markdown highlighting in bare URLs.

So for the above performance improvement we need the empty-region-end hack.
This means we need to match the entire URL in the region-start pattern.
This is slightly challenging because a URL can contain balanced
parentheses but we must not match a trailing ")".
CommonMark [mandates](https://spec.commonmark.org/0.30/#link-destination)
support for at least 3 levels of nested parentheses, so let's just
unroll the regex to implement that.

---

[5MB Markdown file]: https://github.com/mawww/kakoune/issues/4685#issuecomment-1208129806