rc kakrc: add kak=kakrc highlighter alias for markdown/restructuredtext
Filetypes markdown and restructuredtext reuse language highlighting
for code blocks. To highlight a code block of lanugage foo they
essentially do

	require-module foo
	add-highlighter [...] ref foo

This works great if the module name matches the shared
highlighter. This is the case almost all scripts in rc/filetype*.
The only exception is kakrc.kak: the highlighter is named "kakrc"
(just like the filetype) but the module is named "kak".

This requires weird hacks in markdown/restructuredtext.
Ideally we could remove this inconsistency (though that's a breaking
change).  Until we do that, let's add an alias so we can treat
filetypes uniformly.

The following commit will generalize this approach, to allow users
to add arbitrary aliases.