Fix quadratic runtime when updating region highlighter matches
Running %sYeti<ret>casdf on the file
[example.journal.txt](https://github.com/mawww/kakoune/issues/4685#issuecomment-1193243588)
causes noticeable lag.
It creates 6000 selections. Whenever text is inserted/deleted in all
selections, it takes Kakoune 300 milliseconds to update the range
highlighters. This is because the runtime is quadratic in the number
of selections:
for each selection, we call on_new_range(), which calls add_matches()
which calls std::rotate(), which needs needs linear time.

Fix the quadratic behavior by first computing all new ranges and then
adding all matches in one go. This seems to fix the lag.