Allow to traverse selection history tree
Today, selection undo and jump list commands only travel the active
branch of the selection history. Let's make it possible to visit
branches that were undone and abandoned.

I'm not sure how the user interface should look like. For now, let's
make the existing shortcuts traverse the history tree in-order.
This is equivalent to the undo behavior proposed in
https://github.com/mawww/kakoune/pull/4777

THe upside of this UI is that it's dead simple to use because it
needs no additional keys. The potentially annoying thing is that it
requires to press <c-i>/<c-o>/<c-h>/<c-k> more often in some scenarios.