rc grep: -push argument to keep existing buffer
Now that jump-next and friends default to the most recent grep-like
buffer, it would be nice to keep multiple of those buffers, to make
":grep" not destroy state created by previous invocations.

Add a "grep -push" argument that turns on a new "edit -unique" flag
to avoid deleting the existing grep buffer.

Users who choose to use this flag can no longer use the simple *grep*
for returning to the grep buffer. The obvious workaround is to use
the same method that jump-next uses. Add that to the docstring.

{{{

Some discussion:

1. All of grep/lint/make are fifo buffers. Should we close the fifo
   when a new one is created?

2. Is "buffer-latest" the best name?
   Part of the motivation was to avoid breaking
   https://github.com/Delapouite/kakoune-buffers
   though we could definitely arrange that.
   Another argument is that "last" is ambiguous (it could be the most
   recently accessed one).

   I think "latest" fits in reasonably well with the buffer-{next,previous}
   terminology while being less outlandish than "endmost".

   The most promising alternative is "buffer-top", because that fits
   the "stack" mental model. Unfortunately it is ambiguous as well,
   it could mean "first".

   In future "buffer-next", "delete-buffer" and friends should also
   learn a regex argument
}}}