rc grep/jump: default to the last grep-like buffer
I want to run

    grep foo
    jump-next # automagically use the *grep* buffer
    lsp-references
    jump-next # automagically use the *goto* buffer

And have "jump-next" automagically use the right buffer.

To achieve this,
1. Add a new jump_buffers regex option which represents all
   buffers that are used like *grep*. Plugins that create jump buffers
   are meant to add to this global option at plugin-load time.
2. Add a command to find the last buffer that matches a given regex.
3. Make jump-{next,previous} default to the latest buffer that matches
   %opt{jump_buffers}.

Note that this does intentionally not include the *make* buffer
because running :make should not interfere with navigation.

Unfortunately, "edit! -fifo ... *grep*" does not change the position
of the grep buffer in the buflist.  For this feature to work properly,
we need "grep" to move the "*grep*" buffer to the end of the buflist,
else it might be shadowed by other jump buffers.  To make that work,
delete the *grep* buffer to create a fresh one.

The regex matching currently uses POSIX extended regex.  If we
settle on this interface, we should probably fix this.  Either by
providing a kgrep/kakgrep binary in libexec/, or by implementing
"buffer-with-latest" command in C++ outright.

Further reading:
https://discuss.kakoune.com/t/single-command-for-grep-next-match-in-similar-buffers-lsp-make-find/1215
https://lists.sr.ht/~mawww/kakoune/%3C20240225101444.906545-8-aclopte@gmail.com%3E