rc menu: use scratch buffer instead of prompt
The current implementation of menu has scaling problems (it can get
slow or fail due to ARG_MAX). Additionally, menus lack some of the
flexibility and orthogonality of buffers.

I can probably work around the scaling problems by limiting the number
of menu entries. Still, the buffer approach seems worth exploring.

Make menu use a scratch buffer instead of prompt. This drops support
for -select-cmds and -on-abort, which are no longer necessary because
the user can "preview" menu entries by using <ret> and ga in normal
mode.

This includes a number of hacks that we might want to get rid of:
0. It moves the implementation of -auto-single to a separate command.
   Otherwise it would be difficult to avoid ARG_MAX, though that's
   probably not a blocker.
1. When no completion was selected explicitly, insert-mode <ret>
   first injects a <c-n>. This is to support existing menu usage
   scenarios. There is no "-menu" switch (yet?) to mark insert-mode
   completions as authoritative.
2. The command is currently included in the buffer text (and in
   completion).  This keeps things simple but it can be ugly for long
   commands. We might want to hide the command in a line-specs option.
3. The command is run in the context of the buffer that created
   the menu. This works well for use cases that have started relying
   on this (like lsp-code-actions) though we could try to change this.

In future, we can try to also use scratch buffers for composing prompt
commands, like "git edit ".

Closes #5060