Stop recommending static symlink to "%val{runtime}/autoload"
To those people who do not use a Kakoune plugin manager, we recommend
placing extra configuration files in "%val{config}/autoload".

That directory shadows the stdlib at "%val{runtime}/autoload", which
is usually not desired.

To keep loading (parts of) the stdlib, our documentation recommends
creating symlinks along the lines of

	ln -s ${kak_runtime}/autoload ${kak_config}/autoload/stdlib

This workaround is nice in that it's
1. as simple as it gets, and
2. it makes it very obvious that stdlib files are treated just like
   any other plugin; all that matters is the contents of the autoload
   directory.

Unfortunately, this workaround poses a footgun when there are multiple
Kakoune binaries on the same system.

A static symlink can only work for exactly one of src/kak,
/usr/local/bin/kak or /bin/kak. (Third-party plugins tend to be
compatible with a range of versions but the stdlib does not promise
that.)

I think this issue surfaces mostly when switching between versions for
debugging. Probably not very often in practice, but it is too easy
to trigger, because, when creating the autoload symlink you might
not anticipate needing to use a different /path/to/kak.  Similarly,
when debugging an unrelated issue, I usually don't have the presence
of mind to remember that this symlink prevents /bin/kak from working
("HOME=$(mktemp -d) kak" still works but isn't always applicable).

Let's recommend a configuration snippet that will always load the
stdlib of the currently running editor binary.
While at it, make two style changes to the existing autoload logic.

This makes it a little less obvious that stdlib and autoload scripts
are intended to be loadable in any order but this should still be
unambiguous: the added documentation calls out that one can load the
stdlib from either kakrc or user autoload (or kakrc.local I guess).

There are minor risks to documenting this logic; it means that we
cement some decisions like "all autoloadable files must end in .kak"
or "the stdlib does not contain symlinks" (though we could use "find
-L" here too).

In future, we might want to replace the autoload mechanism.
Continuing the direction of this change, we could put the
user in charge of loading both "%val{runtime}/autoload" and
"%val{config}/autoload", to handle stdlib and plugins consistently.
IOW, if, say, "%val{config}/autoload" is a file (not a directory),
do nothing but load that.