Disable recursive mappings when executing user mappings
Commit 8ed16bb2e9 (In non interactive interactive mode, disable user key
mappings, 2014-07-27) made the "noremap" behavior the default, e.g. pressing
x would not trigger the mapping for y here:

	map global normal y :launch-missiles<ret>
	map global normal x y

The same change was not made to the user mode (or custom user modes).
This means that this mapping invokes itself recursively which can cause
a stackoverflow.

	map global user ' ' %{<esc> ...}

Same goes for this mapping (which is how we found this):

	map global user ' ' %{  ...}

Disable mapping resolution in user modes too.