Jujutsu VCS integration
Jujutsu is a VCS with lots of advanced features, and it can be used
alongside Git.

Provide a thin wrapper without much magic. This seems to work well
because jj's command line is already very concise and expressive.

This should make it easy to build a convenient workflow with mappings
such as

	map global jj s %{: jj show } -docstring 'jj show...'
	map global jj i %{:jj split<ret>} -docstring 'jj split'
	map global jj d %{:jj describe } -docstring 'jj describe...'

Perhaps combined with mappings that automagically use the change ID
at cursor[^1] (which are out of scope for this patch).

The notable introduction of magic is ":jj split": because the
first-party "jj split" TUI seems jarring to people who have learned
the ":git apply" style. Make ":jj split" use that style. I have not
yet found a need to use the TUI (I guess splitting commits with file
type changes would be a use case).

~~~

Some commands like "jj show" output to a scratch buffer.

Other commands like "jj new" don't have too much meaningful output,
though displaying the output is still desirable, for learning at
least. Show it in an info box, and log it to the *debug* buffer
(in case the user wants to copy a commit ID after closing the info
box). We could use scratch buffers as well; I'm not yet sure. Maybe
make it configurable.

In future, the same idea could be applied to git.kak.

Note that unlike ":git commit", ":jj describe" keeps its buffer open
after writing, so the change description can be updated multiple times.

~~~

AFAICT there are three commands that require an interactive editor:

1. jj describe (always calls an editor)
2. jj squash (may call an editor once)
2. jj split (always calls a split tool, and may call an editor twice)

Unfortunately all three are handled differently.

1. Our "jj describe" wrapper works like ":git commit": it runs jj twice,
first in a way that intentionally does nothing but extract the
*.jjdescription message template.
The second time we commit the saved message.

2. The above is the most robust one but unfortunately I don't know
how to do the same for "jj squash". We would need a way to make it
do nothing but still output the *.jjdescription message template.
We could pass an editor that always fails, however that would mean:
a. we cannot easily detect other user errors such as "jj squash -r
   no-such-revision" early.
b. we'd need to detect the cases where "jj squash" calls no editor
   (i.e. if at most one of the given revisions has a non-empty
   description).
Let's stick to a single squash invocation; make it async to be able
to use the same editor session for editing the message.  Async is bad
because it prevents scripting but this is new code that we could fix
later if we get a different interface.

3. "jj split" has the same problem as 2, but it currently uses the
(blocking) ":patch" command which invokes "jj split" which invokes
JJ_EDITOR. This prevents a similar async workaround, because the editor
cannot be blocking.  To fix this, we could either add more flexible
interface to jj, or use "patch-range.pl".  For now, simply provide a
fake editor that always preserves the description. I think it should
be fine to change this later, since I only use it interactively.

Note that jj does not yet implement shell tokenization when parsing
JJ_EDITOR; they split by whitespace only. As workaround, create
separate scripts.  Furthermore, the "jj split --tool" argument is
not even split by whitespace so it must be an executable file.

~~~

Regarding prompt completions:
jj provides static and dynamic completions via the clap library.  As of
today, there is no shell-agnostic way of getting completions.  Static
completions as printed by "jj util completion bash" are shell code and
not easily usable by us.  Dynamic completions are more capable, and
require invoking jj each time completions are requested. The output
format is fairly close to the one-candidate-per-line one we need,
so use that.  For my version of jj, this only works when requesting
the fish format -- using "COMPLETE=bash" will produce an empty result.

[^1]: for example, I use this command inside "jj log" output (using
the default jj-log template) to quickly select and capture all change
IDs within my selections.

	define-command -override jj-select-revisions %{
		execute-keys %{<a-s><a-l><semicolon><a-/>^\h*(?:│ )*[@◆○×](?:\h*│)*\h*\b[a-z]+<ret>}
		execute-keys %{1s^\h*(?:│ )*[@◆○×](?:\h*│)*\h*\b([a-z]+)<ret>}
		set-register r %sh{
			eval "set -- $kak_quoted_selections"
			printf %s\\n "$@" | paste -d '|' -s
		}
	}

Then I can type something like

	jj rebase -d trunk() -r <c-r><c-r>

to move the selected commits to a different place in the commit graph.